Checker work
My new thought with the checker is to no longer have it query the api for maps. The methods should take pre-existing beatmap objects, then check those for vlaidity. It can still check leaderboard, but the maps should already exist.
The goal here is to avoid hitting the server multiple times for the same map.