Don't load entire database into memory
Shuffled some things in exporting to google sheets. So instead of converting the database to an array, it gets the sheets-styled row data for each team one at a time, then pushes that rowdata to sheets.
It might end up being slightly slower, but the database shouldn't see any extra traffic as the same query/cursor is being used and iterated on, and I would think memory will be better because only one team object will exist in object form at a time.