New unknown maps added as 'Screenshot Required'
Instead of jumping straight to Pending, add maps with a special status until the player submits a pass for that map.
Maps with the screenshot required status won't show up in the approvers' pending list. (Intended for now, can look at changing it if that turns out to be undesired)

Also fixed some documentation and converted getBeatmap to an async function.