View pending maps
Viewing pending maps works on the new schema now. The output is now split up by mod rather than modpool, and duplicates of the same map shouldn't show with the same mod regardless of pool. Ie a nomod map in the nomod pool and the same nomod map in a custommod pool should only show once under the nomod section.