Export maps
Export maps to google sheets. This function is probably the one I'm not too sure about out of all of them, because of how I wasn't really able to test it properly due to the differences in database structure.
Theoretically I believe it should work once the old teams are cleaned out from the database.