Remove specific maps easier
Should be a bit more forgiving with which modpool to remove from when a mod is given.
CM maps are treated as functionally equivalent to maps in the equivalent modpool, so when specifying a mod by itself if a CM version is before a modpool one in the list that map will be removed first.
I'll probably end up removing from special pools in preference to CM later though.