Add waiting for undo after adding map
Will silently wait for the player to undo any changes when they submit a new map. Nothing will happen if they don't, but if they undo then the pool will be set back to how it was before.
Getting a confirmation is in the process of being moved into a helper function. Currently only the undo command uses it.
Also removed an unused empty function