Add commands modules for bancho and discord
Begin/continue the process of moving commands into a special module for them.
Both disbot.js and boat.js collect command names from the commands modules, and call them by name rather than using if..else.

The commands that are in the new modules seem to work well enough, but the only commands that have been moved are help and check map.