Playfield leniency
Tower of Heaven (847314) has a note one pixel outside the normal playfield, so I've pushed the acceptable area out by one pixel to allow that sort of thing because they're basically close enough.
Also added a timestamp to objects that give issues, so it's easier to locate what's going on with them.