Fix manualadd command and checking member roles
Getting the member associated with a message was broken in the old version of discord.js, updating to the latest version fixed that. By extension checking roles was also broken because members wasn't working.
Discord.js has been updated, so now members can be retreived and roles are now checked properly.