Commands cleanup
Added some updated help messages which need to be loaded in the command constructor
Help now maintains a list of commands so it's not reading from the file system every time the help command is used
Updated validator to TypeScript
Removed 'name' as a CommandArg property. 'arg' is now also the name, and will be treated as such if there isn't a pre-defined way of handling that arg type.