Initial command for adding rules
Basic rules can be specified within the server now. They are added to the database and can be viewed with the view division command.
Only a single week is supported, and there's barely any error checking for bad inputs. I will want to double check how rules are loaded into the checker to make sure changes will be reflected in real time, but that's a problem for later.

Rules list was also moved to the beatmap checker index instead of checker.ts