Make rules week-aware
Adding rules van now be specified for a week-by-week basis. Open division or the first division will be used as a baseline when setting rules for all divisions.
Added 'range' as a known validator type. Also made it so an extended description can be added from within a command even for known validator types.
Ability to specify rank range for a division when the division is created.